Crate opentelemetry_http

Re-exports

pub use bytes::Bytes;

pub use http::Request;

pub use http::Response;

Structs

HeaderExtractor

Helper for extracting headers from HTTP Requests. This is used for OpenTelemetry context propagation over HTTP. See this for example usage.

HeaderInjector

Helper for injecting headers into HTTP Requests. This is used for OpenTelemetry context propagation over HTTP. See this for example usage.

Traits

HttpClient

A minimal interface necessary for sending requests over HTTP. Used primarily for exporting telemetry over HTTP. Also used for fetching sampling strategies for JaegerRemoteSampler

ResponseExt

Methods to make working with responses from the HttpClient trait easier.

Type Aliases

HttpError